pyproject.toml
[tool.poetry]
name = "med-ai-bot-1"
version = "0.1.0"
description = ""
authors = ["Valera386 <valeraorigina@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^20.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
